Move more shared stuff into helper
